+++ cf/data/mixin/deprecations.py
@@ -256,8 +256,8 @@ class DataClassDeprecationsMixin:
         )  # pragma: no cover
 
     def files(self):
-        """Deprecated at version 3.4.0, use method `get_filenames`
-        instead."""
+        """Deprecated at version 3.4.0, consider using method
+        `get_filenames` instead."""
         _DEPRECATION_ERROR_METHOD(
             self,
             "files",
@@ -312,7 +312,11 @@ class DataClassDeprecationsMixin:
     def get_filenames(self):
         """Return the names of files containing parts of the data array.
 
-        Deprecated at version TODODASKVER.
+        Deprecated at version TODODASKVER and and is no longer
+        available. Consider using the `get_original_filenames` method
+        instead.
+
+        .. note:: Might get re-instated in a later version.
 
         :Returns:
 
@@ -321,12 +325,12 @@ class DataClassDeprecationsMixin:
                 data is in memory then an empty `set` is returned.
 
         """
-        return set()
-
-    #        raise DeprecationError(
-    #            "Data method 'get_filenames' has been deprecated at "
-    #            "version TODODASKVER and is not available."
-    #        )  # pragma: no cover
+        _DEPRECATION_ERROR_METHOD(
+            self,
+            "get_filenames",
+            "Consider using the 'get_original_filenames' method instead.",
+            version="TODODASKVER",
+        )  # pragma: no cover
 
     def chunk(self, chunksize=None, total=None, omit_axes=None, pmshape=None):
         """Partition the data array.