+++ Changelog.rst
@@ -3,6 +3,8 @@ version 3.14.0
 
 **202?-??-??**
 
+* New method: `cf.Field.get_original_filenames`
+  (https://github.com/NCAS-CMS/cf-python/issues/448)
 * New method: `cf.Field.to_dask_array`
 * Changed dependency: ``1.10.0.1<=cfdm<1.10.1.0``
 * New dependency: ``dask>=2022.6.0``